AB#26873 - Zones - Refactor Application Contracts
