--- applications/Unity.GrantManager/src/Unity.GrantManager.Application.Contracts/SettingManagement/GetZoneListResultDto.cs
@@ -1,7 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Unity.GrantManager.SettingManagement;
-public class GetZoneListResultDto
-{
-    public List<ZoneGroupDefinitionDto> ZoneGroups { get; set; }
-}
