+++ applications/Unity.GrantManager/src/Unity.GrantManager.Application.Contracts/Zones/ZoneDefinitionDto.cs
@@ -1,13 +1,12 @@
-﻿using System.Collections.Generic;
+﻿namespace Unity.GrantManager.Zones;
 
-namespace Unity.GrantManager.SettingManagement;
-
-public class ZoneTabDefinitionDto
+public class ZoneDefinitionDto
 {
     public string Name { get; set; } = string.Empty;
     public string? DisplayName { get; set; }
     public bool IsEnabled { get; set; }
+    public string ViewComponentType { get; set; } = string.Empty;
     public int SortOrder { get; set; } = 1000;
     public string? ElementId { get; set; }
-    public List<ZoneDefinitionDto> Zones { get; set; } = [];
+    public object? Arguments { get; set; }
 }
