+++ applications/Unity.GrantManager/src/Unity.GrantManager.Application.Contracts/Zones/ZoneTabDefinitionDto.cs
@@ -1,12 +1,13 @@
-﻿namespace Unity.GrantManager.SettingManagement;
+﻿using System.Collections.Generic;
 
-public class ZoneDefinitionDto
+namespace Unity.GrantManager.Zones;
+
+public class ZoneTabDefinitionDto
 {
     public string Name { get; set; } = string.Empty;
     public string? DisplayName { get; set; }
     public bool IsEnabled { get; set; }
-    public string ViewComponentType { get; set; } = string.Empty;
     public int SortOrder { get; set; } = 1000;
     public string? ElementId { get; set; }
-    public object? Arguments { get; set; }
+    public List<ZoneDefinitionDto> Zones { get; set; } = [];
 }
